services/scraper/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
motor==3.6.0
selenium==4.23.1